% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_caseCounts.R
\name{create_caseCounts}
\alias{create_caseCounts}
\title{Title}
\usage{
create_caseCounts(date_vec, location_vec, cases_vec)
}
\arguments{
\item{date_vec}{}

\item{location_vec}{}

\item{cases_vec}{}
}
\description{
Title
}
