% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_linelist.R
\name{create_linelist}
\alias{create_linelist}
\title{Title}
\usage{
create_linelist(
  caseCounts,
  reportF = NULL,
  reportF_args = NULL,
  reportF_missP = NULL
)
}
\arguments{
\item{caseCounts}{}

\item{reportF}{}

\item{reportF_args}{}

\item{reportF_missP}{}
}
\description{
Title
}
