\name{linelistBayes-package}
\alias{linelistBayes-package}
\alias{linelistBayes}
\docType{package}
\title{
\packageTitle{linelistBayes}
}
\description{
\packageDescription{linelistBayes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linelistBayes}
\packageIndices{linelistBayes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{linelistBayes}

Maintainer: \packageMaintainer{linelistBayes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
