% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_backnow.R
\name{run_backnow}
\alias{run_backnow}
\title{Title}
\usage{
run_backnow(
  input,
  MAX_ITER = 2000,
  norm_sigma = NULL,
  sip = NULL,
  NB_maxdelay = NULL,
  NB_size = NULL,
  workerID = NULL,
  printProgress = 0,
  ...
)
}
\arguments{
\item{input}{}

\item{MAX_ITER}{}

\item{norm_sigma}{}

\item{sip}{}

\item{NB_maxdelay}{}

\item{NB_size}{}

\item{workerID}{}

\item{printProgress}{}

\item{...}{}
}
\description{
Title
}
