% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_cases}
\alias{sample_cases}
\alias{sample_dates}
\alias{sample_location}
\alias{sample_onset_dates}
\alias{sample_report_dates}
\alias{out_list_demo}
\title{Sample datasets}
\format{
Either vectors or a list object (out_list_demo)
}
\usage{
sample_cases

sample_dates

sample_location

sample_onset_dates

sample_report_dates

out_list_demo
}
\description{
These data sets provide tests for use in converting between case counts
and line list data.
}
\details{
`sample_cases` provides a vector of daily case counts.
`sample_dates` are the dates of the sample case counts.
`sample_location` are the locations of the sample case counts.
`sample_onset_dates` are the same information as in sample_cases, but 
with one entry per case indicating the date of symptom onset.
`sample_report_dates` are the same information as in sample_cases, but 
with one entry per case indicating the date of symptom reporting.
`out_list_demo` is a precomputed output from run_backnow, useful for
plotting in the vignettes
}
\examples{
sample_cases
sample_dates
sample_location
sample_onset_dates
sample_report_dates
}
\keyword{datasets}
