% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.R
\name{si}
\alias{si}
\title{Title}
\usage{
si(ndays, alpha, beta)
}
\arguments{
\item{ndays}{}

\item{alpha}{}

\item{beta}{}
}
\description{
There are different types of serial interval used here
 https://www.sciencedirect.com/science/article/pii/S1201971220306111
 p[j] is the probability of a serial interval of j days, exactly, and
 not the sum of all things before it, so this gives the PDF, not the CDF
 right because pgamma() gives the CDF, not the PDF
 so this just creates the PDF
 so a vector of length 14 with alpha and beta as defined in the doc, for COVID-19
}
